fix warnings

+++ Pod/Main.cpp
@@ -120,7 +120,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine
 		}
 		else {
 			appController->Handler_Keyboard(hWnd);
-			if (GetTickCount() - startTime > (1000 / setting->getFPS())) {
+			if (GetTickCount() - (DWORD)startTime > (DWORD)(1000 / setting->getFPS())) {
 				appController->Game_Run(hWnd, Setting::getInstance()->getDt());
 				startTime = GetTickCount();
 			}
